update packages and dependencies to eliminate vulnerabilities
Some of the packages we use with Hedwig were outdated and contained
dependencies that had vulnerabilities ranging from minor to major.
By updating to the latest versions, and removing or replacing some, we
close many of these issues, all while staying current.

Some small changes to the code had to be done to adapt it to the newer
versions of some packages, and to remove both errors and warnings from
the logs when building or running.

Notable changes:
- babel-cli, babel-preset-env and babel-preset-es2015 updated to
  @babel/cli and @babel/preset-env
- live-reload changed to livereload
- postcss-cssnext updated to postcss-preset-env
- rollup-watch is now included in rollup
- Minus can not preced lonely variables in calc(), changed to "* -1"
- Extending at-rules cannot occur within unnamed rules
- Updated livereload script-tag according to new package